requirements.txt
pydantic>=2.7,<3
rank-bm25==0.2.2
sentence-transformers==2.7.0
pypdf==4.2.0
pyyaml==6.0.2
tqdm==4.66.4
numpy>=1.26
